+++ site/src/components/Resources/AgentRow.tsx
@@ -35,7 +35,7 @@ import {
   WorkspaceAgentMetadata,
 } from "../../api/typesGenerated"
 import { AppLink } from "./AppLink/AppLink"
-import { SSHButton } from "../SSHButton/SSHButton"
+import { SSHButton } from "./SSHButton/SSHButton"
 import { Stack } from "../Stack/Stack"
 import { TerminalLink } from "./TerminalLink/TerminalLink"
 import { AgentLatency } from "./AgentLatency"
